+++ src/pudl/metadata/codes.py
@@ -1,6 +1,10 @@
 """Metadata for cleaning, re-encoding, and documenting coded data columns.
 
-These dictionaries are used to create Encoder instances. They contain the following keys:
+These dictionaries are used to create Encoder instances. Each key is a table name with
+a sub dictionary that includes additional detail. The table names must end with the
+data_source as a sufix (for EIA 860, 861 or 923 tables include ``_eia``).
+
+The table-specific dictionaries contain the following keys:
 
 * 'df': A dataframe associating short codes with long descriptions and other information.
   Each dataframe needs at least three standard columns: "code", "label", "description".
@@ -230,6 +234,21 @@ CODE_METADATA: dict[str, dict[str, Any]] = {
             "Wholesale Power Marketer": "W",
         },
         "ignored_codes": [],
+    },
+    "boiler_generator_association_types_eia": {
+        "df": pd.DataFrame(
+            columns=[
+                "code",
+                "label",
+                "description",
+            ],
+            data=[
+                ("A", "actual", "An actual boiler generator association."),
+                ("T", "theoretical", "A theoretical boiler generator association."),
+            ],
+        ),
+        "code_fixes": {"t": "T", "a": "A"},
+        "ignored_codes": ["1"],
     },
     "operational_status_eia": {
         "df": pd.DataFrame(
@@ -1109,6 +1128,59 @@ CODE_METADATA: dict[str, dict[str, Any]] = {
                     7,
                     "industrial_cogen",
                     "Industrial cogeneration facilities that produce electric power, are connected to the grid, and can sell power to the public",
+                ),
+            ],
+        ).convert_dtypes(),
+        "code_fixes": {},
+        "ignored_codes": [],
+    },
+    "steam_plant_types_eia": {
+        "df": pd.DataFrame(
+            columns=["code", "label", "description"],
+            data=[
+                (
+                    1,
+                    "steam_over_100mw",
+                    "Plants with combustible-fueled steam-electric generators with a sum of 100 MW or more steam-electric nameplate capacity (including combined cycle steam-electric generators with duct firing).",
+                ),
+                (
+                    2,
+                    "steam_between_10mw_and_100mw",
+                    "Plants with combustible-fueled steam-electric generators with a sum of 10 MW or more but less than 100 MW steam-electric nameplate capacity (including combined cycle steam-electric generators with duct firing).",
+                ),
+                (
+                    3,
+                    "nuclear_over_100mw",
+                    "Plants with nuclear fueled generators, combined cycle steam-electric generators without duct firing and solar thermal electric generators using a steam cycle with a sum of 100 MW or more steam-electric nameplate capacity.",
+                ),
+                (
+                    4,
+                    "non_steam",
+                    "Plants with non-steam fueled electric generators (wind, PV, geothermal, fuel cell, combustion turbines, IC engines, etc.) and electric generators not meeting conditions of categories above.",
+                ),
+            ],
+        ).convert_dtypes(),
+        "code_fixes": {},
+        "ignored_codes": [],
+    },
+    "reporting_frequencies_eia": {
+        "df": pd.DataFrame(
+            columns=["code", "label", "description"],
+            data=[
+                (
+                    "A",
+                    "annual",
+                    "The respondent only provides an annual total(s) for this record via the EIA-923 annual survey form.  Any monthly data in this record is estimated based on the respondent's reported annual total(s) and power plants with similar characteristics to this plant.",
+                ),
+                (
+                    "M",
+                    "monthly",
+                    "The respondent provides monthly values for this record and does so via the EIA-923 monthly survey form.",
+                ),
+                (
+                    "AM",
+                    "monthly_annual",
+                    "The respondent provides monthly values for this record, but does so once per year via the EIA-923 annual survey form.",
                 ),
             ],
         ).convert_dtypes(),